+++ libopendmi/include/opendmi/entity/additional-info.h
@@ -11,37 +11,31 @@
 
 #include <opendmi/entity.h>
 
-typedef struct dmi_additional_info_data dmi_additional_info_data_t;
-typedef struct dmi_additional_info      dmi_additional_info_t;
+typedef struct dmi_additional_info       dmi_additional_info_t;
+typedef struct dmi_additional_info_entry dmi_additional_info_entry_t;
 
 /**
  * @brief Additional information entry.
  */
-dmi_packed_struct(dmi_additional_info_entry_data)
+struct dmi_additional_info_entry
 {
-    /**
-     * @brief Length of this additional information entry instance; a minimum
-     * of 6.
-     */
-    dmi_byte_t length;
-
     /**
      * @brief Handle, or instance number, associated with the structure for
      * which additional information is provided.
      */
-    dmi_word_t ref_handle;
+    dmi_handle_t ref_handle;
 
     /**
      * @brief Offset of the field within the structure referenced by the
      * referenced handle for which additional information is provided.
      */
-    dmi_byte_t ref_offset;
+    unsigned ref_offset;
 
     /**
      * @brief Number of the optional string to be associated with the field
      * referenced by the referenced offset.
      */
-    dmi_string_t string;
+    const char *string;
 
     /**
      * @brief Enumerated value or updated field content that has not yet been
@@ -52,23 +46,29 @@ dmi_packed_struct(dmi_additional_info_entry_data)
      * This field is the same type and size as the field being referenced by
      * this additional information entry.
      */
-    dmi_data_t value[];
+    dmi_data_t value[32];
+
+    /**
+     * @brief Length of the value data (determined from additional info
+     * structure, must match length of the referenced field).
+     */
+    size_t value_length;
 };
 
 /**
  * @brief Additional information structure (type 40).
  */
-dmi_packed_struct(dmi_additional_info_data)
+struct dmi_additional_info
 {
-    /**
-     * @brief SMBIOS structure header.
-     */
-    dmi_header_t header;
-
     /**
      * @brief Number of additional information entries that follow.
      */
-    dmi_byte_t entry_count;
+    size_t entry_count;
+
+    /**
+     * @brief Additional information entries.
+     */
+    dmi_additional_info_entry_t *entries;
 };
 
 /**